add AddViewModel
